% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrees.R
\name{agrees}
\alias{agrees}
\title{Simply create tests to evaluate a value}
\usage{
agrees(expr, message = "")
}
\arguments{
\item{expr}{an unquoted expression written in terms of 'x' that returns a TRUE/FALSE value.}

\item{message}{an optional failure message}
}
\description{
Turn a logical expression into a test for check_() functions.
}
