% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_is.R
\name{arg_is}
\alias{arg_is}
\title{Simple interface to argument checking}
\usage{
arg_is(value = NULL, hint = FALSE, test = NULL)
}
\arguments{
\item{value}{-- a simple value that can be tested by equality. If \code{test} is also specified,
the value will be used only in a hint, in which case it should be set to something meaningful
as a hint.}

\item{hint}{if TRUE, give a more diagnostic version of the source of any error}

\item{test}{-- a function that carries out the test,
e.g. match_number(), check_class(), check_data_frame()}
}
\value{
a character string message. Length zero means the test was passed.
}
\description{
Use \code{arg_is()} as an argument to \code{check_argument()}
}
\examples{
\dontrun{
 test <- check_argument("grab_this + 1", arg_is(3))
}

}

