% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{before}
\alias{before}
\title{glue to hold the test statements together}
\usage{
before(EX)
}
\arguments{
\item{EX}{A set of expressions of the sort returned by capture.code()}
}
\description{
glue to hold the test statements together
}

