% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture.code}
\alias{capture.code}
\title{evaluate R statements, capturing code, created vars, and printed output}
\usage{
capture.code(text = NULL, dot_R_file = NULL, example = NULL)
}
\arguments{
\item{dot_R_file}{the name of the file containing the code as text}

\item{example}{the name of the example to run (this is just for testing and should
be replaced by a \code{text = } argument for production), which is an alternative
to \code{dot_R_file}.}
}
\value{
a list containing 1) the \code{results} of each command (which might not have been
saved in an object in the statement itself), 2) a list \code{envlist}
holding the environments created for each command, 3) the statements themselves as
character strings, one string for each command regardless of how many lines
it was spread over in the code, and 4) the statements as expressions (which might
be evaluated in checking code, or parsed, or whatever).
}
\description{
Each expression in the code is evaluated in it's own environment, with the
previous expression's environment as the parent. That way, you can examine
what's changed after each statement.
}
\details{
searching for objects in the last environment will move as needed to the
parent environments, and so will give the most recent value of each object.
}
\examples{
foo <- capture.code(example = "example_0.R")
}

