% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_argument.R
\name{check_argument}
\alias{check_argument}
\title{Apply a test to the value of an argument}
\usage{
check_argument(arg_spec, test)
}
\arguments{
\item{arg_spec}{character string describing what we want to match, e.g. "sin(grab_this)"}

\item{test}{is a function that takes the concordance produced by <arg_spec>
and returns a message saying what's wrong. If that message is "", the test passes.
Works on a single line}
}
\value{
a capture object
}
\description{
Apply a test to the value of an argument
}

