% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_info_to_file.R
\name{check_info_to_file}
\alias{check_info_to_file}
\title{Save the checking information to a file}
\usage{
check_info_to_file(label = NULL, user_code = NULL, check_code = NULL,
  envir_result = NULL, evaluate_result = NULL, ...)
}
\arguments{
\item{label}{info handed over by tutor}

\item{user_code}{same}

\item{check_code}{same}

\item{envir_result}{same}

\item{evaluate_result}{same}

\item{other}{possible arguments}
}
\description{
Saves to a file the checking information tutor generates when a code-exercise
is submitted. Once saved, the information can be used for building and debugging
check code in the usual development environment.
}

