% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{check_model_formula}
\alias{check_model_formula}
\title{Check components of models}
\usage{
check_model_formula(soln_formula, message = "Give a real message")
}
\arguments{
\item{soln_formula}{a formula that the model should have}

\item{df}{a data frame that the model should have been fitted to}
}
\description{
Try to find a model object that matches the formula
}

