% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tutor_tests.R
\name{checkr_run_tests}
\alias{checkr_run_tests}
\title{Run the tests from info produced by tutor}
\usage{
checkr_run_tests(label = NULL, user_code = NULL, check_code = NULL,
  envir_result = NULL, evaluate_result = NULL, ...)
}
\arguments{
\item{tutor_obj}{the checking object produced by tutor
You can also use the result of check_info_from_file()}
}
\description{
Run the tests from info produced by tutor
}

