% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{checkr_tutor}
\alias{checkr_tutor}
\title{Interface to tutor}
\usage{
checkr_tutor(label = NULL, user_code = NULL, solution_code = NULL,
  check_code = NULL, envir_result = NULL, evaluate_result = NULL, ...,
  debug = FALSE)
}
\arguments{
\item{label}{character string containing name of the chunk that's being checked. For \code{run_tests_from_file()}, this
will be used to refer to the saved file.}

\item{user_code}{character string with the student's code}

\item{solution_code}{character string with the code provided by the question's author.}

\item{check_code}{the character string contents of the chunk that specifies the
checks to run on the chunk being checked}

\item{envir_result}{the R environment after the student's code has been run
NOTE NOTE NOTE I would like to have the R environment *before* the student code is run. That way,
I can run the solution_code in the same environment seen by the student code.}

\item{evaluate_result}{the output of evaluate::evaluate from the code input. Tutor provides
this, but checkr doesn't use it. IT MAY BE DESIRABLE TO USE IT IN THE FUTURE.}

\item{...}{other arguments (see tutor documentation)}

\item{debug}{if \code{TRUE}, write the information in the other arguments to a file
\code{"~/Downloads/CheckR/chunk-[chunk-name].rds"}.
Then, the stored information can be put through
the tests outside of the tutor system, allowing better debugging. See \code{run_tests_from_file()} which
reads the RDS file and runs the tests.}
}
\value{
a list in the format required by tutor
}
\description{
The function that the \code{tutor} document will call when checking
a submission. You can tell \code{tutor} to do this with a statement in the \code{tutor}
document like \code{tutor_options(exercise.checker = checkr::checkr_tutor)}.
}
\details{
The arguments are set by the tutor system. Only \code{debug} isn't
standard.
}

