% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_to.R
\name{closest_to}
\alias{closest_to}
\title{Locators based on solution code}
\usage{
closest_to(what)
}
\arguments{
\item{what}{either an object name or an identifying character string from the code generating
the object}
}
\value{
a function that takes a capture as input and returns a capture as output
}
\description{
\code{closest_to()} finds a matching line in the solution code. It then uses the
value of that line (in the solution code) to look for the best-matching
value in the student code. The function is not to be used stand-alone, but
rather in the \code{\link{soln_test}} function, which calls it once for the solution
code and again for the user code. For the solution code, it returns the matching line
but for the user code it returns the line whose value matches the most closely the
value of the line in the solution code. So long as there is some
}

