% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_tests.R
\name{any_test}
\alias{all_tests}
\alias{any_test}
\alias{branch_test}
\alias{compose_tests}
\title{Combine multiple tests into a single test}
\usage{
any_test(test_1, test_2, ...)

all_tests(...)

branch_test(condition, yes, no)

compose_tests(test_1, test_2, ...)
}
\arguments{
\item{test_1}{a locator test}

\item{test_2}{another locator test}

\item{...}{optionally, more locator tests}

\item{condition}{a test to determine which of \code{yes} or \code{no}}

\item{yes}{the test to run if \code{condition} passes}

\item{no}{the test to run if \code{condition} fails}
}
\description{
\code{any_test(t1, t2, ...)} If any of the tests pass, the whole set of tests
passes.
}
\details{
\code{all_tests(t1, t2, ...)} All of the given tests must pass for the test to pass.

\code{branch_test(condition, yes, no)}

\code{compose_tests(t1, t2, ...)} combine a list of tests into a test sequence.
}

