% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{check_info_to_file}
\alias{check_info_to_file}
\alias{check_info_from_file}
\alias{run_tests_from_file}
\title{Functions for debugging checkr}
\usage{
check_info_to_file(label = NULL, user_code = NULL, check_code = NULL,
  envir_result = NULL, evaluate_result = NULL, ...)

check_info_from_file(label, directory = "~/Downloads/CheckR/")

run_tests_from_file(label)
}
\arguments{
\item{label}{character string name of the problem chunk. This is info handed over by learnr.}

\item{user_code}{same}

\item{check_code}{same}

\item{envir_result}{same}

\item{evaluate_result}{same}

\item{...}{other possible arguments}

\item{directory}{which directory to save the file in}
}
\description{
The basic idea is to save the information generated by learnr generates when a code-exercise
is submitted. Once saved in a file, the information can be used for building and debugging
check code in the usual development environment.
}
