% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/either.R
\name{either}
\alias{either}
\title{Combine results of locator tests}
\usage{
either(test_1, test_2, ...)
}
\arguments{
\item{test_1}{a locator test}

\item{test_2}{another locator test}

\item{...}{optionally, more locator tests}
}
\description{
\code{either()} works on multiple tests. If any of the tests pass, the whole set of tests
passes.
}

