% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\name{expand_chain}
\alias{expand_chain}
\title{Expand a Magrittr chain so arguments resolve correctly}
\usage{
expand_chain(chain_string)
}
\arguments{
\item{chain_string}{a text string containing a Magrittr chain}
}
\description{
Expand a Magrittr chain so arguments resolve correctly
}

