% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{f_has_terms}
\alias{f_has_terms}
\title{checks on formula's}
\usage{
f_has_terms(student, term_formula)
}
\arguments{
\item{student}{the student's formula}

\item{term_formula}{the reference formula}
}
\description{
checks on formula's
}

