% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcall.R
\name{fcall}
\alias{fcall}
\title{Find matching lines based on function calls.}
\usage{
fcall(fun_spec, message = NULL, mistake = FALSE, hint = FALSE)
}
\arguments{
\item{fun_spec}{character string describing the function to look for
and whatever args are to be matched, e.g. "2 + 2".
Use \code{whatever} for any arguments for which
a place is needed but for which you don't care what the value is.
"success" means that the pattern was found.}

\item{message}{a character string message to be returned on failure. If
not provided, a suitable message will be constructed based on \code{fun_spec}.}

\item{mistake}{if \code{TRUE}, then finding a matching line causes a failure. This is
useful for testing for common mistakes.}

\item{hint}{if \code{TRUE}, signals downstream functions to provide a more diagnostic
failure message. (Not used at present.)}
}
\value{
a function that takes a capture object as input and returns
a capture argument noting the line that matches (if any).
}
\description{
A functional form for testing functions and arguments
with this, you can set up the test beforehand and pass the code to it later.
}

