% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigns_to.R
\name{find_assignment}
\alias{find_assignment}
\title{Locator test for assignment to a specific name or regex}
\usage{
find_assignment(name = NULL, message = NULL, hint = FALSE, regex = NULL)
}
\arguments{
\item{name}{a character string or unquoted bare name that should be the name assigned to the result of a code line.}

\item{message}{character string with the failure message}

\item{hint}{if \code{TRUE}, failure message should show the name to the student}

\item{regex}{a regex that defines the match}
}
\value{
a function that can be used in a locator test sequence. That function
takes a capture as input and returns a capture as output
}
\description{
This locator test looks for an assignment to the named variable or,
if \code{regex} is given, to a name matching the regex.
}
\details{
You can specify both \code{name} and \code{regex}. If both are specified, the
failure message will refer to \code{name} (if \code{hint = TRUE}).
}
