% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.R
\name{find_content}
\alias{find_content}
\title{look for matches in the code}
\usage{
find_content(where = c("returns", "names", "statements", "commands"), what,
  mode = c("class", "number", "match", "regex"),
  message = "Test writer: give a meaningful fail message", mistake = FALSE,
  ...)
}
\arguments{
\item{where}{character string specifying whether to look at names of objects
assigned names, objects
produced by the commands (and potentially not named),
the statements themselves. NOT SURE WHAT ROLE commands will play.  Those are the
expressions as expressions.}

\item{what}{character string, regex, class name, number range, or name assigned
that we're looking for}

\item{mode}{what kind of thing to search for.}

\item{mistake}{to indicate that, to pass,
the pattern should not be found. Can be used to check for common
mistakes.}
}
\value{
A list with 3 elements: \code{passed} indicating TRUE/FALSE whether the
test was passed; \code{line} the line number where the item was found; \code{code}
the same as the input.
}
\description{
Just the first such line is reported.
}
\details{
Look for the line, if any, containing the sought-for object.
}

