% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_formula.R
\name{has_formula}
\alias{has_formula}
\title{Look for a formula in a command line and apply a test to the formula}
\usage{
has_formula(test = NULL, message = "couldn't find a formula")
}
\arguments{
\item{test}{a test like \code{match_formula()}. Default: Does the formula exist?}

\item{message}{character string message to give on failure}
}
\description{
Look for a formula in a command line and apply a test to the formula
}
\seealso{
\code{\link{check_value}}, \code{\link{check_argument}}
}

