% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.R
\name{in_factory}
\alias{in_factory}
\title{Locator functions}
\usage{
in_factory(where)
}
\description{
Look for the line in the code matching the identified text, function call, etc.
}
\details{
# These are factories for test functions. Test functions take either a result or the
}

