% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_statements.R
\name{in_statements}
\alias{in_statements}
\title{Look for matches to the contents of the statements themselves}
\usage{
in_statements(pattern, message = NULL, regex = FALSE, mistake = FALSE)
}
\arguments{
\item{pattern}{character string containing the pattern to search for in the statements}

\item{message}{character string giving the failure message}

\item{regex}{if \code{TRUE} treat the pattern as a regex}

\item{mistake}{if \code{TRUE}, the test fails if the pattern is located}
}
\description{
Look for matches to the contents of the statements themselves
}

