% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_google.R
\name{set_up_google_auth_token}
\alias{set_up_google_auth_token}
\alias{turn_on_google_logging}
\title{Functions for logging checkr results to a Google Spreadsheet}
\usage{
set_up_google_auth_token(run = FALSE)

turn_on_google_logging(sheet_name)
}
\arguments{
\item{run}{set this to TRUE if you really want to re-authenticate.
Run this only once, before uploading your app to the server
It sets up the authentication for Google}

\item{sheet_name}{character string name of the Google sheet that will be used for logging.}
}
\description{
Functions for logging checkr results to a Google Spreadsheet
}

