% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_for.R
\name{look_for}
\alias{look_for}
\title{Look for some given content in the printed form of the objects created}
\usage{
look_for(what = NULL, description = " ", from_code, go_in_reverse = FALSE,
  where = "results")
}
\arguments{
\item{what}{a character string. The string is treated as a regex.}

\item{description}{a character string describing to the student what is being sought}

\item{go_in_reverse}{(logical) the order in which to assess the results}

\item{results}{the object produced by capture()}
}
\description{
Redo this so that look_for() et al. return objects which can be operated on by
and, or, etc.
}

