% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_values.R
\name{match_values}
\alias{match_values}
\title{supervisory function for matching output values}
\usage{
match_values(student = NULL, solution = NULL, ...)
}
\arguments{
\item{solution}{the evaluated capture object from the solution code}

\item{...}{lists of statements giving names to the objects to be matched
and the requirements framed on the objects created in \code{assigns}}

\item{capture}{the evaluated capture object from the submitted code}
}
\description{
supervisory function for matching output values
}
\details{
the elements in \code{...} have two forms. Named elements are set to statements
that identify a line in the code. The returned value of that line will be given the name.
Unnamed elements consist of comparison tests (see \code{same_()}) that operate on the
objects given names in previous elements of the list.
}

