% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_passing.R
\name{set_success_message}
\alias{get_success_message}
\alias{set_success_message}
\title{message passing from tests}
\usage{
set_success_message(message)

get_success_message()
}
\arguments{
\item{message}{a character string message to signal success to the user.}
}
\description{
for communicating back from the tests to the final result.
}

