% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\docType{data}
\name{my_test}
\alias{my_test}
\title{Expand a Magrittr chain so arguments resolve correctly}
\format{An object of class \code{character} of length 1.}
\usage{
my_test
}
\description{
Expand a Magrittr chain so arguments resolve correctly
}
\keyword{datasets}

