% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_objects.R
\name{named_object}
\alias{named_object}
\title{Look within the objects created and assigned by the code}
\usage{
named_object(what, description, from_code, go_in_reverse = FALSE, caps = I)
}
\arguments{
\item{what}{A regex of the object name}

\item{description}{a for-the-user description of the object}

\item{from_code}{the code object created by capture()}

\item{go_in_reverse}{whether to look forward or backward}

\item{caps}{What to do about capitalization. Default: nothing. You can also set it to be
\code{tolower()} or \code{toupper()} as appropriate to match your regex}
}
\description{
Look within the objects created and assigned by the code
}

