% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{needs}
\alias{needs}
\title{modifiers on pass objects}
\usage{
needs(pass)
}
\arguments{
\item{pass}{a pass object}

\item{capture}{a capture object}
}
\value{
the modified pass object
}
\description{
These functions let the test-writer specify whether it's sufficient that a test be passed
and whether it's required (\code{needs()}) that a test pass.
}

