% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{find_pipe_start}
\alias{check_pipe_input}
\alias{check_pipe_output}
\alias{find_pipe_start}
\title{Locator and check functions relating to pipes}
\usage{
find_pipe_start(pattern = NULL, message = "no \%>\% pipe found.")

check_pipe_input(test, message = "")

check_pipe_output(test, message = "")
}
\arguments{
\item{pattern}{a character string or complete expression starting the pipe.}

\item{message}{character string failure message}
}
\description{
Locator and check functions relating to pipes
}

