% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-evaluation-check.R
\name{check_function_calls}
\alias{check_assignment_names}
\alias{check_function_calls}
\title{Pre-evaluation checking of code}
\usage{
check_function_calls(user_code, fnames)

check_assignment_names(user_code)
}
\arguments{
\item{user_code}{a character string containing the user's code}

\item{fnames}{a character string vector of function names to check}
}
\value{
a character string. All pre-check functions return a character string (for now).
If the string has length 0, the check passes. Non-zero length strings should be
diagnostic messages or hints for users.
}
\description{
These functions support checking of user code before it is evaluated. This is
useful when the built in parsing error messages from R are not gentle enough for
beginning users.
}
\details{
\code{check_function_calls()} checks whether functions are called with opening and closing parentheses.
You need to provide a list of the functions it should look for (since no parens are needed for
object names).

\code{check_assignment_names()} checks whether assignments are made to a legal name (or other
syntactically legal construction).
}

