% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{Run the tests from info produced by tutor}
\usage{
run_tests(label = NULL, user_code = NULL, check_code = NULL,
  envir_result = NULL, evaluate_result = NULL, ...,
  feedback = tutor::feedback, debug = TRUE)
}
\arguments{
\item{tutor_obj}{the checking object produced by tutor
You can also use the result of check_info_from_file()}
}
\description{
Run the tests from info produced by tutor
}

