% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.R
\name{same_}
\alias{is.test_result}
\alias{same_}
\alias{same_num}
\alias{same_vec}
\title{Compare submitted code values to solution code values}
\usage{
same_(compare_fun, objname, hint = FALSE, ...)

same_vec(objname, hint = FALSE, ...)

same_num(objname, hint = FALSE, ...)

is.test_result(x)
}
\arguments{
\item{compare_fun}{the function to compare two objects, e.g. compare_numbers, compare_classes}

\item{objname}{the name of the object. Should be an unquoted name (e.g. \code{b}),
not a character string (e.g., not \code{"b"}). Expressions
are also allowed, e.g. \code{sqrt(b)}}

\item{hint}{Set to \code{TRUE} to provide more diagnostic messages to the student.}

\item{...}{additional parameters to describe the match}

\item{x}{the object to be tested}
}
\value{
a list of with class \code{"test_result"}. See \code{new_test_result()}.
}
\description{
Or, more precisely, create functions that will compare the values
of identically named objects in two environments or lists. Typically,
one of the environments will be from the submitted answers, the other will
be from the official solution.
}
\details{
#' \code{same_num()} compares numerical vectors, providing a tolerance for numerical
disagreement, \code{same_vec()} compares vectors for exact equality.

\code{same_} lets the test writer apply her own comparison function. The others are for
convenience for common operations use
built in comparison functions.

There is a distinction between these \code{same_()} functions and comparison
functions. A comparison function is an ordinary function that takes three required arguments. \code{S} and
\code{R} are the objects to be compared. \code{hint} is a boolean flag whether to make
the failure message more diagnostic than the default. Comparison functions return \code{""} to
indicate a pass, and a string of non-zero length to indicate a failure (and hopefully provide
some guidance to the student.)

In contrast, the \code{same_()} functions apply comparison functions to the corresponding
values from the submitted code and the solution (or reference) code.
}

