% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.R
\name{same_}
\alias{same_}
\title{Compare pairs of values}
\usage{
same_(compare_fun, name, hint = FALSE, ...)
}
\arguments{
\item{compare_fun}{function to compare two objects, e.g. compare_numbers, compare_classes}

\item{name}{the name of the object. Should be an unquoted name (e.g. \code{b}),
not a character string (e.g., not \code{"b"}). Expressions
are also allowed, e.g. \code{sqrt(b)}}

\item{...}{additional parameters to describe the match}
}
\description{
Or, more precisely, create functions that will compare the values
of identically named objects in two environments or lists. Typically,
one of the environments will be from the submitted answers, the other will
be from the official solution.
}

