% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soln_test.R
\name{soln_test}
\alias{soln_test}
\title{supervisory function testing whether student code matches a given solution}
\usage{
soln_test(student = NULL, solution = NULL, ...)
}
\arguments{
\item{student}{the evaluated capture object from the submitted code}

\item{solution}{the evaluated capture object from the solution code}

\item{...}{lists of statements giving names to the objects to be matched
and the requirements framed on the objects created in \code{assigns}}
}
\description{
supervisory function testing whether student code matches a given solution
}
\details{
the elements in \code{...} have two forms. Named elements are set to statements
that identify a line in the code. The returned value of that line will be given the name.
Unnamed elements consist of comparison tests (see \code{same_()}) that operate on the
objects given names in previous elements of the list.
}
