% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{then}
\alias{then}
\title{glue to hold the test statements together}
\usage{
then(capture)
}
\arguments{
\item{capture}{A set of expressions of the sort returned by capture.code()}
}
\description{
These operate on the output of a locator test.
\code{then} restricts the subsequent tests to the lines after the one identified
\code{previously} restricts the subsequent tests to the lines before the one identified
}

