% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{val_by_name}
\alias{val_by_name}
\title{Checking values of objects created or displayed by the code}
\usage{
val_by_name(what = what, message = "Give a real message", strict = TRUE)
}
\arguments{
\item{what}{the object name, class(es), function/argument string, or text to match}

\item{strict}{whether the matching name or text can be approximate
(e.g. capitalization, dots, underscores, ...)}

\item{capture}{the code object}
}
\description{
These functions capture the object itself. If there is no such object, a character
string with the message is returned. That message will have a 'no_match' attribute
}
\details{
If the test passes, the value is returned in the "value" attribute of the capture object
}

