% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\docType{data}
\name{check_value}
\alias{check_value}
\alias{grab_this}
\alias{whatever}
\title{Checking values of objects created or displayed by the code}
\format{An object of class \code{name} of length 1.}
\usage{
check_value(test, message = NULL, mistake = FALSE)

grab_this

whatever
}
\arguments{
\item{test}{a function taking one value as an argument. The test to
run to evaluate the captured value. Test functions should
return \code{""} if passing, non-empty message string if not.}

\item{message}{the message to give if the test fails}

\item{mistake}{if \code{TRUE}, the specified test is one that should NOT be seen in
the code.}
}
\value{
a function that will take a capture object as input and return
a capture object as output. The value of the line (identified by previous locator tests)
will be compared to the reference value given when constructing the test.
}
\description{
\code{check_value} applies a test to the value produced by a line of code. The tests
themselves are "match functions" such as \code{match_number}, \code{match_vector}, \code{match_names},
\code{match_formula}, \code{match_data.frame} that do the comparison.
}
\details{
These functions capture the object itself. If there is no such object, a character
string with the message is returned. That message will have a 'no_match' attribute
If the test passes, the value is returned in the "value" attribute of the capture object
}
\seealso{
\code{\link{check_argument}}
}
\keyword{datasets}
