% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{check_value}
\alias{check_value}
\title{Checking values of objects created or displayed by the code}
\usage{
check_value(test, message = NULL, mistake = FALSE)
}
\arguments{
\item{test}{the test to run to evaluate the captured value}

\item{message}{the message to give if the test fails}

\item{capture}{a capture object}
}
\description{
These functions capture the object itself. If there is no such object, a character
string with the message is returned. That message will have a 'no_match' attribute
}
\details{
If the test passes, the value is returned in the "value" attribute of the capture object
}

