% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R, R/values.R
\name{check_value}
\alias{check_value}
\title{Check values of objects}
\usage{
check_value(capture, test, message = "Give a helpful message!")

check_value(capture, test, message = "Give a helpful message!")
}
\arguments{
\item{capture}{a capture object}

\item{test}{a function that takes one argument, the value to be checked.
Typically this function will be constructed by functions such as the test to run to evaluate the captured value}

\item{message}{the message to give if the test fails}

\item{capture}{a capture object}

\item{test}{the test to run to evaluate the captured value}

\item{message}{the message to give if the test fails}
}
\description{
Check values of objects
}

