% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/either.R
\name{any_test}
\alias{all_tests}
\alias{any_test}
\alias{branch_test}
\title{Combine results of locator tests}
\usage{
any_test(test_1, test_2, ...)

all_tests(...)

branch_test(condition, yes, no)
}
\arguments{
\item{test_1}{a locator test}

\item{test_2}{another locator test}

\item{...}{optionally, more locator tests}
}
\description{
These functions combine multiple tests into a single test.
}
\details{
\code{any_test(t1, t2, ...)} If any of the tests pass, the whole set of tests
passes.

\code{all_tests(t1, t2, ...)} All of the given tests must pass for the test to pass.

\code{branch_test(condition, yes, no)}
}

