% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locator_functions.R
\name{has_constants}
\alias{find_statement}
\alias{has_constants}
\alias{has_names}
\alias{in_values}
\alias{similar_names}
\title{Locator functions}
\usage{
has_constants(..., message = NULL, mistake = FALSE)

has_names(pattern, message = NULL, mistake = FALSE)

find_statement(pattern, message = NULL, regex = FALSE, mistake = FALSE)

similar_names(x, message = NULL)

in_values(x, message = NULL)
}
\arguments{
\item{...}{placeholder for future possible arguments}

\item{message}{character string giving the failure message}

\item{mistake}{if \code{TRUE}, the test fails if the pattern is located}

\item{pattern}{character string containing the pattern to search for in the statements}

\item{regex}{if \code{TRUE} treat the pattern as a regex}

\item{x}{the target that you are looking for in the values.
Specify the match in terms of a comparison function}
}
\description{
\code{find_statement} looks for character matches to the code
}
\details{
\code{in_values} looks for a match to the values produced by the code

\code{has_names} looks for variables and functions in the expression that match all of those
given in \code{pattern}. The \code{pattern} should be an unquoted expression. A good
trick if you just want a set of names, e.g. \code{fred}, \code{betty}, \code{sarah}, is to
set pattern to a kind of function call: \code{fred(betty, sarah)}.

\code{has_constants} looks for numbers and character strings in an expression. (These are different from function names
and object names.)

\code{similar_names} takes a character string containing an R expression.
It looks for the statement in the submitted code that has a set of names
that most closely matches those in the character string. It never fails, always
returning a capture object pointing to some statement. However, the heuristic
used for comparison is just a heuristic!
}
\examples{
U <- capture.code("ifelse(sin(37 + 14) > .5, 'yes', 'no way')")
test_1 <- has_constants(37, 14)

}
\seealso{
other locator functions \code{\link{find_assignment}}, \code{\link{find_call}}
}

