% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R, R/message_passing.R
\name{authentication}
\alias{authentication}
\alias{get_success_message}
\alias{get_user_name}
\alias{set_success_message}
\alias{set_user_name}
\title{Authenticate a user}
\usage{
authentication(password_file, ...)

set_success_message(message)

get_success_message()

set_user_name(user_name)

get_user_name()
}
\arguments{
\item{password_file}{The name of a file containing user-ID/password pairs to use in the
authentication. This is intended for future use.}

\item{...}{other arguments for future use.}

\item{message}{a character string message to signal success to the user.}

\item{user_name}{a character string identifying the user.}
}
\description{
Use this function to insert into a tutor document the user input fields for
user ID and password. This is done in a simple, but unusual way: by creating an empty chunk with
a `child` chunk attribute like this: \code{child = checkr::authentication()}.

for communicating back from the tests to the final result.
}
\details{
\code{set_user_name} and \code{get_user_name} are used by the logging system
}

