% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{find_pipe_start}
\alias{find_pipe_start}
\alias{check_pipe_input}
\alias{check_pipe_output}
\title{Locator and check functions relating to pipes}
\usage{
find_pipe_start(pattern = NULL, message = "no \%>\% pipe found.")

check_pipe_input(test, message = "")

check_pipe_output(test, message = "")
}
\arguments{
\item{pattern}{a character string or complete expression starting the pipe.}

\item{message}{character string failure message}

\item{test}{a test function to run for find_pipe_input() or find_pipe_output()}
}
\description{
Locator and check functions relating to pipes
}
